+++ main.js
@@ -24,8 +24,6 @@ function createWindow() {
   mainWindow.on("closed", function () {
     mainWindow = null;
   });
-
-  mainWindow.webContents.openDevTools();
 }
 
 function startServer() {
